Create MetricsFactory to provide a centralized registry for metrics
This change extracts the Prometheus registry initialization code from Astra.java
into a new MetricsFactory class that provides global access to the metrics registry.
Similar to logging implementations, this allows components across the application
to access a shared meter registry without having to pass it through constructors.

Benefits:
- Provides static access to metrics registry similar to the logger pattern
- Ensures consistent configuration of the registry across the application
- Centralizes tag management for better metrics organization
- Simplifies instantiation of meter objects in new components
